Refactor string and arbitrary class names
